move service proxy itself
